Implement a more typed system to prevent issues
The way the actions currently work is that they magically assign some
stuff to the context and the Koa framework takes care of the rest. Magic
is bad and can lead to confusion and issues later down the line. For
instance, a hanging action could be difficult to debug, where as simple
action that either returns a promise or throws an error should be simple
to follow.

Error handling is also added to catch these exceptions as they come.