Split the logger setup
We'd like to be using the logger outside the app context. To achieve
that, we can separate the middleware into two functions:

"setupLogger" - responsible for getting the instance out
"logger" - taking the current responsibility of being a middleware and
attaching it to the system

We're also fixing the service name in the logger for the errors to be
represented properly.